chore(grasshopper): refactor common unwrapping logic
